+++ ios/CaptureSdk-Bridging-Header.h
@@ -1,0 +1,10 @@
+#import <React/RCTBridgeModule.h>
+#import <React/RCTEventEmitter.h>
+#import <ReactCommon/RCTTurboModule.h>
+#import <ReactCommon/TurboModule.h>
+#import <React/RCTUtils.h>
+#import <React/RCTViewManager.h>
+#import <React/RCTUIManager.h>
+#import <CaptureSDK/CaptureSDK.h>
+#import "SktEvent.h"
+#import "CaptureSdkSpec.h"